+++ mod-source-record-manager-server/src/test/java/org/folio/services/ChangeEngineServiceImplTest.java
@@ -1,14 +1,13 @@
 package org.folio.services;
 
 import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.containsString;
 import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.hasSize;
-import static org.hamcrest.Matchers.notNullValue;
 import static org.hamcrest.Matchers.nullValue;
 import static org.junit.Assert.assertTrue;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyList;
+import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.when;
 
 import java.util.Collections;
@@ -21,6 +20,7 @@ import io.vertx.core.Future;
 import io.vertx.core.Vertx;
 import io.vertx.core.json.Json;
 import io.vertx.core.json.JsonObject;
+import org.folio.rest.jaxrs.model.MappingMetadataDto;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -66,14 +66,20 @@ public class ChangeEngineServiceImplTest {
   private RecordsPublishingService recordsPublishingService;
   @Mock
   private KafkaConfig kafkaConfig;
+  @Mock
+  private MappingMetadataService mappingMetadataService;
   private OkapiConnectionParams okapiConnectionParams = new OkapiConnectionParams(new HashMap<>(), Vertx.vertx());
 
   @InjectMocks
   private ChangeEngineServiceImpl service;
 
   @Before
-  public void setUp() throws Exception {
+  public void setUp() {
     ReflectionTestUtils.setField(service, "maxDistributionNum", 10);
+    ReflectionTestUtils.setField(service, "batchSize", 100);
+
+    when(mappingMetadataService.getMappingMetadataDto(anyString(), any(OkapiConnectionParams.class)))
+      .thenReturn(Future.succeededFuture(new MappingMetadataDto()));
   }
 
   @Test
@@ -123,7 +129,7 @@ public class ChangeEngineServiceImplTest {
   }
 
   @Test
-  public void shouldReturnMarcHoldingsRecordWithErrorsWhen004FieldIsMissing() {
+  public void shouldNotReturnMarcHoldingsRecordWhen004FieldIsMissing() {
     RawRecordsDto rawRecordsDto = getTestRawRecordsDto(MARC_HOLDINGS_REC_WITHOUT_004);
     JobExecution jobExecution = getTestJobExecution();
 
@@ -135,11 +141,7 @@ public class ChangeEngineServiceImplTest {
     Future<List<Record>> serviceFuture = executeWithKafkaMock(rawRecordsDto, jobExecution, Future.succeededFuture(true));
 
     var actual = serviceFuture.result();
-    assertThat(actual, hasSize(1));
-    assertThat(actual.get(0).getRecordType(), equalTo(Record.RecordType.MARC_HOLDING));
-    assertThat(actual.get(0).getErrorRecord(), notNullValue());
-    assertThat(actual.get(0).getErrorRecord().getDescription(),
-      containsString("The 004 tag of the Holdings doesn't has a link"));
+    assertThat(actual, hasSize(0));
   }
 
   @Test